requirements.txt
flask
flask-cors
sqlalchemy
psycopg2-binary
python-dotenv
streamlit
requests
pytest
google-genai
pytest